Extract SurfaceHost from TerminalControl
TerminalControl carried three jobs in one 1165-line class: owning
the ghostty surface (attach / detach order, swap chain binding,
input translation, IME), rendering the pane's overlays (badges,
scrollbar, search bar), and deciding which of those two gets the
keyboard. The search-box pty leak in #171 came from the second and
third being tangled with the first in the same focus handlers.

Move the first job into SurfaceHost, a plain (non-XAML) class the
control owns through a shared_ptr and hands its SwapChainPanel to.
The control keeps XAML focus, ProtectedCursor, the overlays and the
ISurfaceView facade; every routed event is forwarded to the host,
and the host consults one injected predicate (SetInputGate) before
letting keystrokes, IME engagement or IME commits reach the pty —
the composite answers it with "the search box does not hold focus".

The public Attach / Detach / Rehost / Surface() surface stays on
the control as forwarders so Tab, TabFactory and MainWindow are
untouched apart from the swap-chain callback registration moving
to SurfaceHost. The _DEBUG SearchLeak traces from the #171
investigation are dropped with the code they instrumented.
